Minor improvements to make SqlQueueTrigger discoverable by the functions runtime
- Change default message type to string (instead of object). Full-fledged support for different binding types is coming later.
- Make SqlQueueExtensionConfig a public class so that the functions runtime can discover it.
- Decorate SqlQueueTriggerAttribute and SqlQueueAttribute with the [Binding] attribute so that the functions runtime can discover it
- To have this extension picked up by the functions runtime, create a directory 'extensions', and set environment variable 'AzureWebJobs_ExtensionsPath' to point to it. Create a sub-directory in the 'extensions' directory and copy the directory Microsoft.Azure.WebJobs.Extensions.dll to it.
