+++ src/ExtensionsSample/Samples/SqlQueueSamples.cs
@@ -11,13 +11,13 @@ namespace ExtensionsSample
 {
     public static class SqlQueueSamples
     {
-        public static void SqlQueue_Trigger_Basic([SqlQueueTrigger(DatabaseName = "webjobs", QueueName = "RECEIVERQUEUE1", ContractName = "C", MessageDataType = "VarChar(Max)")] object queueMessage)
+        public static void SqlQueue_Trigger_Basic([SqlQueueTrigger(DatabaseName = "webjobs", QueueName = "RECEIVERQUEUE1", ContractName = "C", MessageDataType = "VarChar(Max)")] string queueMessage)
         {
             Console.WriteLine("Message1 : " + queueMessage);
         }
 
         // Note: Param name "messageValue" has to match the metadata generated by SqlQueueTrigger
-        public static void SqlQueue_Trigger_With_Metadata([SqlQueueTrigger( DatabaseName = "webjobs", QueueName = "RECEIVERQUEUE2", ContractName = "C")] object queueMessage, SqlQueueTriggerValue messageValue)
+        public static void SqlQueue_Trigger_With_Metadata([SqlQueueTrigger(QueueName = "RECEIVERQUEUE2", ContractName = "C")] string queueMessage, SqlQueueTriggerValue messageValue)
         {
             Console.WriteLine("Message2 : " + messageValue);
         }
