+++ src/WebJobs.Extensions.SqlQueue/SqlQueueExtensionConfig.cs
@@ -1,0 +1,26 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+using System;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Extensions.SqlQueue;
+using Microsoft.Azure.WebJobs.Host.Config;
+
+namespace Microsoft.Azure.WebJobs
+{
+    public class SqlQueueExtensionConfig : IExtensionConfigProvider
+    {
+        public void Initialize(ExtensionConfigContext context)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            // Register our extension binding providers
+            context.Config.RegisterBindingExtensions(
+                new SqlQueueAttributeBindingProvider(),
+                new SqlQueueTriggerAttributeBindingProvider());
+        }
+    }
+}
